vulnerable-app/requirements.txt
flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
openai==1.3.0
python-dotenv==1.0.0
langchain==0.0.292
pydantic==2.4.2 